python-uv-django/pyproject.toml
[project]
name = "python-uv-django"
version = "0.1.0"
description = "A modern Django 5.0+ template using uv for package management with Django REST Framework"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "django>=5.0.0,<6.0.0",
    "djangorestframework>=3.14.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "dj-database-url>=2.1.0",
    "django-cors-headers>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-watch>=4.2.0",
    "pytest-django>=4.7.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "pylint>=3.0.0",
    "pylint-django>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
indent-width = 2

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "DJ",   # flake8-django
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"*/migrations/*" = ["E501", "DJ01"]  # Allow long lines and nullable strings in migrations
"*/settings/*" = ["F405"]  # Allow star imports in settings

[tool.ruff.lint.isort]
known-first-party = ["src", "config", "apps"]
section-order = ["future", "standard-library", "third-party", "django", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
django = ["django", "rest_framework"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ["py312"]
skip-string-normalization = true
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | migrations
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--reuse-db",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/migrations/*",
    "*/manage.py",
    "*/asgi.py",
    "*/wsgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings.development"

[[tool.mypy.overrides]]
module = [
    "factory.*",
    "faker.*",
]
ignore_missing_imports = true

[tool.pylint.'MASTER']
load-plugins = ['pylint_django']

[tool.pylint.format]
max-line-length = 100
indent-string = '  '  # 2 spaces

[tool.pylint.messages_control]
disable = [
  'C0103',  # invalid-name
  'C0114',  # missing-module-docstring
  'C0115',  # missing-class-docstring
  'C0116',  # missing-function-docstring
  'R0903',  # too-few-public-methods
  'W0212',  # protected-access
  'C0330',  # bad-continuation
  'C0326',  # bad-whitespace
  'E1101',  # no-member (pylint-django handles this)
]

[tool.pylint.basic]
good-names = ['i', 'j', 'k', 'ex', 'id', 'db', 'pk', '_']
